# docs/swagger.yaml
basePath: /
definitions:
  e.BackStruct:
    properties:
      code:
        type: integer
      data:
        type: string
      msg:
        type: string
    type: object
host: 192.168.50.237
info:
  contact:
    email: support@swagger.io
    name: 这里写联系人信息
    url: '[http://www.swagger.io/support](http://www.swagger.io/support)'
  description: 测试用程序
  license:
    name: Apache 2.0
    url: '[http://www.apache.org/licenses/LICENSE-2.0.html](http://www.apache.org/licenses/LICENSE-2.0.html)'
  termsOfService: '[http://swagger.io/terms/](http://swagger.io/terms/)'
  title: go_project学习项目
  version: "1.0"
paths:
  /api/v1/articles:
    get:
      parameters:
      - description: State
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取多个文章标签
    post:
      parameters:
      - description: tag_id
        in: query
        name: tag_id
        type: integer
      - description: title
        in: query
        name: title
        type: string
      - description: desc
        in: query
        name: desc
        type: string
      - description: content
        in: query
        name: content
        type: string
      - description: State
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增文章
  /api/v1/articles/:id:
    delete:
      parameters:
      - description: id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 删除文章
    get:
      parameters:
      - description: id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取单个文章
    put:
      parameters:
      - description: id
        in: query
        name: id
        type: integer
      - description: tag_id
        in: query
        name: tag_id
        type: integer
      - description: title
        in: query
        name: title
        type: string
      - description: desc
        in: query
        name: desc
        type: string
      - description: content
        in: query
        name: content
        type: string
      - description: modified_by
        in: query
        name: modified_by
        type: string
      - description: State
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 修改文章
  /api/v1/tags:
    get:
      parameters:
      - description: Name
        in: query
        name: name
        required: true
        type: string
      - description: State
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取多个文章标签
    post:
      parameters:
      - description: Name
        in: query
        name: name
        required: true
        type: string
      - description: State
        in: query
        name: state
        type: integer
      - description: CreatedBy
        in: query
        name: created_by
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增文章标签
  /api/v1/tags/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 删除文章标签
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: ID
        in: query
        name: name
        required: true
        type: string
      - description: State
        in: query
        name: state
        type: integer
      - description: ModifiedBy
        in: query
        name: modified_by
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 修改文章标签
  /auth:
    get:
      parameters:
      - description: username
        in: query
        name: username
        type: string
      - description: password
        in: query
        name: password
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 登录
swagger: "2.0"